Numbered images and matching descriptions
Now all of the images and corresponding images have been numbered. Also, I included an excel file that shows where the images are included in the survey: Description/Pretest/Test. Also, they are seperated into two folders as well. The numbering system and excel sheet will hopefully make it easier to work with the data after collecting it.