Reduced number of items in study
I have brought it down to only 10 descriptions and 10 images in the pre test. Also made it so that none of the descriptions have corresponding images in the pre-test. The post test has 25 images, some of which were only described in the pre-test, some which were shown in the pre-test, and some that did not show up at all previously in the study.